Check

+++ NhaMapThep.Domain/Entities/ConfigTable/MucSanPhamEntity.cs
@@ -14,6 +14,6 @@ namespace NhaMapThep.Domain.Entities.ConfigTable
         public required int MucSanPhamToiThieu { get; set; }
         public required int MucSanPhamToiDa { get; set; }
         [Column(TypeName = "decimal")]
-        public required double LuongMucSanPham { get; set; }
+        public required decimal LuongMucSanPham { get; set; }
     }
 }
